Move txn summary view
