+++ components/cards/ClaimSummaryView.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import type { SidebarProps } from "../Sidebar";
 
-type ClaimViewProps = Pick<SidebarProps, "isFairLaunch"> & {
+type ClaimSummaryViewProps = Pick<SidebarProps, "isFairLaunch"> & {
   /** equal to the Network Fee rate divided by 31,536,000 (seconds in a 365 day year). */
   stream: string;
   /** equal to (`flowRate` - `oldFlowRate`) * 2 * 60 * 60. */
@@ -13,11 +13,11 @@ type ClaimViewProps = Pick<SidebarProps, "isFairLaunch"> & {
  * and we should use FairClaimLaunchAction,
  * else use SimpleClaimAction
  */
-function ClaimView({
+function ClaimSummaryView({
   stream,
   streamBuffer,
   isFairLaunch = false,
-}: ClaimViewProps) {
+}: ClaimSummaryViewProps) {
   return (
     <div>
       <h3>Transaction Summary (ETHx)</h3>
@@ -28,4 +28,4 @@ function ClaimView({
   );
 }
 
-export default ClaimView;
+export default ClaimSummaryView;
